requirements.txt
# Install necessary python3 packages
pillow==8.1.2
mysql-connector-python==8.0.23
mysqlclient==2.0.3
requests==2.25.1
pywebpush==1.11.0
websockets==9.1
python-memcached==1.59
trml2pdf==0.6
markdown==3.3.
uritemplate==3.0.1
pygments==2.8.1
unipath==1.1
django-rest-passwordreset==1.3.0

# Install django related packages
django==3.2.15
django-import-export==2.5.0
django-user-agents==0.4.0
django-currencies==0.10.1

django-tracking-analyzer==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==5.2.2
django-filter==2.4.0
django-guardian==2.3.0

dj-database-url==0.5.0
icmplib==3.0.3
ajaxuploader
